complete loading ui for home,restaurants and menu details
