+++ src/components/skeletons/MenuSkeleton.jsx
@@ -1,0 +1,10 @@
+export default function MenuSkeleton() {
+  return (
+    <div>
+      <span className='loading h-56 rounded-lg'></span>
+      <span className='loading h-6 rounded-lg w-36 mt-12'></span>
+      <span className='loading h-6 rounded-lg w-28 mt-3 mb-5'></span>
+      <span className='loading h-6 rounded-lg w-20'></span>
+    </div>
+  );
+}
