--- src/components/skeletons/RestaurantCard.jsx
@@ -1,9 +1,0 @@
-export default function RestaurantCard() {
-  return (
-    <div className='space-y-2'>
-      <span className='loading h-40'></span>
-      <span className='loading h-5 w-1/2'></span>
-      <span className='loading h-5 w-1/2'></span>
-    </div>
-  );
-}
