+++ src/components/skeletons/RestaurantSkeleton.jsx
@@ -1,13 +1,12 @@
-export default function HomeSkeleton() {
+export default function RestaurantSkeleton() {
   const items = Array.from({ length: 5 }, (_, index) => index);
   return (
-    <div className='space-y-4'>
+    <div className='space-y-8'>
       <span className='loading h-10 rounded-lg'></span>
-      <span className='loading h-7 rounded-lg'></span>
 
       <div className='grid gap-4 '>
         {items.map((item) => (
-          <span key={item} className='loading h-48 rounded-2xl'></span>
+          <span key={item} className='loading h-36 rounded-2xl'></span>
         ))}
       </div>
     </div>
