+++ src/index.css
@@ -63,7 +63,7 @@ button {
 
 .loading {
   position: relative;
-  background: #ebedf5;
+  background: #f7f7f7;
   border-radius: 0.5rem;
   overflow: hidden;
   display: block;
